Extract grpc statuses to separate module for reuse
CHANGELOG_BEGIN
CHANGELOG_END
