+++ ledger/participant-state/BUILD.bazel
@@ -33,6 +33,7 @@ da_scala_library(
         "//ledger/ledger-configuration",
         "//ledger/ledger-configuration/protobuf:ledger_configuration_proto_java",
         "//ledger/ledger-offset",
+        "//ledger/ledger-grpc",
         "//ledger/metrics",
         "//libs-scala/logging-entries",
         "@maven//:com_google_protobuf_protobuf_java",
